@@ -1,8 +1,9 @@
 package ru.mtsbank.entity;
 
+import jakarta.persistence.*;
 import lombok.Data;
 
-import javax.persistence.*;
+
 import java.util.List;
 
 @Entity
